feat(buildozer): add support for formatting on write using the -buildifier flag
This patch adds support to the `buildozer` rules for using the
`-buildifier` flag to format the BUILD files that it operates on. Users
are able to specify a new attribute, `format_on_write=True` in order to
set the `-buildifier` flag to an empty value, which uses buildozer's
internal formatter.

If the `buildifier` attribute is also set, we pass that executable's
path in as the value to `-buildifier`, which causes `buildozer` to use
that executable instead of its internal formatter.